add __slots__ to providers
