+++ that_depends/providers/__init__.py
@@ -1,3 +1,4 @@
+from that_depends.providers.attr_getter import AttrGetter
 from that_depends.providers.base import AbstractProvider, AbstractResource
 from that_depends.providers.collections import Dict, List
 from that_depends.providers.context_resources import (
@@ -18,6 +19,7 @@ __all__ = [
     "AsyncContextResource",
     "AsyncFactory",
     "AsyncResource",
+    "AttrGetter",
     "ContextResource",
     "DIContextMiddleware",
     "Factory",
