+++ that_depends/providers/singleton.py
@@ -1,6 +1,7 @@
 import asyncio
 import typing
 
+from that_depends.providers import AttrGetter
 from that_depends.providers.base import AbstractProvider
 
 
@@ -8,19 +9,9 @@ T = typing.TypeVar("T")
 P = typing.ParamSpec("P")
 
 
-class AttrGetter(AbstractProvider[T]):
-    def __init__(self, provider: AbstractProvider[T], attr_name: str) -> None:
-        self._provider = provider
-        self._attr_name = attr_name
-
-    async def async_resolve(self) -> typing.Any:  # noqa: ANN401
-        return getattr(await self._provider.async_resolve(), self._attr_name)
-
-    def sync_resolve(self) -> typing.Any:  # noqa: ANN401
-        return getattr(self._provider.sync_resolve(), self._attr_name)
-
-
 class Singleton(AbstractProvider[T]):
+    __slots__ = "_factory", "_args", "_kwargs", "_override", "_instance", "_resolving_lock"
+
     def __init__(self, factory: type[T] | typing.Callable[P, T], *args: P.args, **kwargs: P.kwargs) -> None:
         self._factory = factory
         self._args = args
